+++ src/modules/f1/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default as f1Routes } from './routes.js';