+++ src/modules/gallery/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default as galleryRoutes } from './routes.js';