+++ src/modules/medical-journal/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default as medJournalRoutes } from './routes.js';